[FE][S13P21D108-127] fix: 환전 예약 로직 오류 수정

Closes S13P21D108-127

See merge request s13-fintech-finance-sub1/S13P21D108!162

+++ frontend/app/src/main/java/com/d108/moyeo/domain/repository/ExchangeRepository.kt
@@ -12,7 +12,7 @@ interface ExchangeRepository {
         fromCurrency: String,
         toCurrency: String,
         amount: Long,
-        targetRate: Double,
+        targetRate: Long,
         expiresAt: String
     ): Result<Unit>
